Merge latent state profile matrix (H_star) and Q matrix

This function

merges identical rows in H: pseudo-cluster ---> scientific clusters;

merges partner latent states (identical non-zero columns in H); merge the corresponding rows in Q by pmax. Even though identifiability condition may prevent non-identical columns in H, because we specify a maximum latent state dimension m_max, MCMC may produce identical columns in some iterations.

merge_H_Q(
  H_star_redun,
  mylist,
  t,
  Q,
  VERBOSE = FALSE,
  z_pseudo = NULL,
  skip_Q = FALSE
)

Arguments

H_star_redun
A binary matrix of rows t_max+3 by m_max; It may be redundant because it has rows corresponding to empty pseudo-clusters, and if the matrix has at least one non-zero element, the zero columns indicate inactive latent states

mylist
Cluster labels; A vector of length t_max+3 comprised of integers that need not be consecutive.

t
The number of non-empty pseudo-clusters

Q
The m_max by L Q matrix

VERBOSE
Default to FALSE: no print of the merged matrices; Otherwise set to TRUE.

z_pseudo
Default is NULL. A vector of pseudo-cluster indicators to be merged into scientific cluster indicators z_sci.

skip_Q
Default is FALSE - merge the rows of Q that correspond to partner states; Set to TRUE if Q is given in which case only rows of H_star_redun will be merged (keep the zero columns in H_star_redun, and not merging partner states).

Value

A list comprised of two elements named: H_star_merge and Q_merge and z_sci if z_pseudo is provided.

Examples

# the third latent state is inactive among non-empty pseudo-clusters;
# the 2nd and 4th latent states are partners.
# merge 1st and 2nd pseudo-clusters.
 H_star_redun <- matrix(c(0,1,0,1,
                          0,1,0,1,
                          1,0,0,0,
                          0,0,0,0,
                          0,0,0,0,
                          0,0,0,0,
                          0,0,0,0),nrow=7,ncol=4,byrow=TRUE)
mylist <- c(1,2,3,5,0,0,0)
t <- 4
Q <- simulate_Q(4,100,p=0.1)
z_pseudo <- c(1,1,3,5,2,2,2,1,3,3,5,1,2,2,1)
# Q is unknown and sampled
merge_H_Q(H_star_redun,mylist,t,Q,TRUE,z_pseudo)
#> >> ignored 1 `non-active` latent states ---> 3 latent states. 
#> >> absorbed 1 pseudo clusters ---> 3 scientific clusters.
#> >> absorbed 1 `partner` latent states ---> 2 latent states. 
#> $H_star_merge
#>      010 100
#> [1,]   0   1
#> [2,]   1   0
#> [3,]   0   0
#> 
#> $Q_merge
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14]
#> [1,]    0    0    0    0    0    0    1    0    0     0     0     0     1     0
#> [2,]    1    0    0    0    0    0    0    0    1     0     1     0     0     0
#>      [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25] [,26]
#> [1,]     0     0     0     0     0     0     0     0     0     0     1     0
#> [2,]     0     0     0     0     1     0     0     0     0     0     0     1
#>      [,27] [,28] [,29] [,30] [,31] [,32] [,33] [,34] [,35] [,36] [,37] [,38]
#> [1,]     0     0     0     0     0     0     0     0     0     1     1     0
#> [2,]     0     0     1     0     0     1     1     1     1     0     0     0
#>      [,39] [,40] [,41] [,42] [,43] [,44] [,45] [,46] [,47] [,48] [,49] [,50]
#> [1,]     0     0     0     1     0     0     0     0     0     0     1     0
#> [2,]     0     0     0     0     1     0     0     1     0     1     1     0
#>      [,51] [,52] [,53] [,54] [,55] [,56] [,57] [,58] [,59] [,60] [,61] [,62]
#> [1,]     0     0     1     0     0     0     0     1     0     1     0     0
#> [2,]     0     1     1     0     0     0     0     1     0     0     1     0
#>      [,63] [,64] [,65] [,66] [,67] [,68] [,69] [,70] [,71] [,72] [,73] [,74]
#> [1,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [2,]     1     0     0     0     0     0     0     0     0     1     1     1
#>      [,75] [,76] [,77] [,78] [,79] [,80] [,81] [,82] [,83] [,84] [,85] [,86]
#> [1,]     1     0     0     0     0     0     0     0     0     0     0     0
#> [2,]     0     1     0     0     0     0     0     0     0     0     0     0
#>      [,87] [,88] [,89] [,90] [,91] [,92] [,93] [,94] [,95] [,96] [,97] [,98]
#> [1,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [2,]     1     0     0     0     0     0     1     0     0     1     0     0
#>      [,99] [,100]
#> [1,]     0      0
#> [2,]     0      0
#> 
#> $z_sci
#>  [1] 1 1 2 3 1 1 1 1 2 2 3 1 1 1 1
#> 
# Q is known:
merge_H_Q(H_star_redun,mylist,t,Q,TRUE,z_pseudo,TRUE)
#> >> absorbed 1 pseudo clusters ---> 3 scientific clusters.
#> >> absorbed 1 `partner` latent states ---> 3 latent states. 
#> $H_star_merge
#>      [,1] [,2] [,3] [,4]
#> 0101    0    1    0    1
#> 1000    1    0    0    0
#> 0000    0    0    0    0
#> 
#> $Q_merge
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14]
#> [1,]    0    0    0    0    0    0    1    0    0     0     0     0     1     0
#> [2,]    1    0    0    0    0    0    0    0    1     0     1     0     0     0
#> [3,]    0    0    0    0    0    0    0    1    0     0     0     0     0     0
#> [4,]    0    0    0    0    0    0    0    0    0     0     0     0     0     0
#>      [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25] [,26]
#> [1,]     0     0     0     0     0     0     0     0     0     0     1     0
#> [2,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [3,]     0     0     0     0     0     0     0     0     0     0     0     1
#> [4,]     0     0     0     0     1     0     0     0     0     0     0     1
#>      [,27] [,28] [,29] [,30] [,31] [,32] [,33] [,34] [,35] [,36] [,37] [,38]
#> [1,]     0     0     0     0     0     0     0     0     0     1     1     0
#> [2,]     0     0     0     0     0     0     1     1     1     0     0     0
#> [3,]     0     0     0     0     0     0     0     0     0     0     0     1
#> [4,]     0     0     1     0     0     1     0     0     0     0     0     0
#>      [,39] [,40] [,41] [,42] [,43] [,44] [,45] [,46] [,47] [,48] [,49] [,50]
#> [1,]     0     0     0     1     0     0     0     0     0     0     1     0
#> [2,]     0     0     0     0     1     0     0     1     0     0     1     0
#> [3,]     0     0     0     0     0     0     0     0     0     0     0     1
#> [4,]     0     0     0     0     0     0     0     0     0     1     0     0
#>      [,51] [,52] [,53] [,54] [,55] [,56] [,57] [,58] [,59] [,60] [,61] [,62]
#> [1,]     0     0     1     0     0     0     0     1     0     1     0     0
#> [2,]     0     1     0     0     0     0     0     1     0     0     1     0
#> [3,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [4,]     0     0     1     0     0     0     0     0     0     0     0     0
#>      [,63] [,64] [,65] [,66] [,67] [,68] [,69] [,70] [,71] [,72] [,73] [,74]
#> [1,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [2,]     1     0     0     0     0     0     0     0     0     0     1     0
#> [3,]     0     0     0     0     1     0     1     0     1     0     0     0
#> [4,]     0     0     0     0     0     0     0     0     0     1     0     1
#>      [,75] [,76] [,77] [,78] [,79] [,80] [,81] [,82] [,83] [,84] [,85] [,86]
#> [1,]     1     0     0     0     0     0     0     0     0     0     0     0
#> [2,]     0     1     0     0     0     0     0     0     0     0     0     0
#> [3,]     0     0     0     0     1     0     0     0     0     0     0     1
#> [4,]     0     0     0     0     0     0     0     0     0     0     0     0
#>      [,87] [,88] [,89] [,90] [,91] [,92] [,93] [,94] [,95] [,96] [,97] [,98]
#> [1,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [2,]     1     0     0     0     0     0     1     0     0     1     0     0
#> [3,]     0     0     0     0     1     0     1     1     0     0     0     1
#> [4,]     0     0     0     0     0     0     0     0     0     1     0     0
#>      [,99] [,100]
#> [1,]     0      0
#> [2,]     0      0
#> [3,]     0      0
#> [4,]     0      0
#> 
#> $z_sci
#>  [1] 1 1 2 3 1 1 1 1 2 2 3 1 1 1 1
#>