Randomly generate a Q matrix within the identifying condition (data-driven)

This function initializes Q (if unknown) during MCMC sampling chain within identifiability constraints. It is a warm start - because it will not assign an "1" to dimension "l" with few ones in the data. NB: harder to get 1 to zero? easy to get zero to one?

simulate_Q_dat(M, dat, p = 0.1, frac = 1/4)

Arguments

M
latent state dimension

dat
binary data matrix (rows for observations, columns for dimensions)

p
Bernoulli probability of 1 in the Q matrix (except two diagonal matrices)

frac
A threshold - this function only initializes the dimensions with at least frac*100 percent observed frequencies.

Value

a binary matrix of dimension M by L

Examples

# simulate data:
L0 <- 100
options_sim0  <- list(N = 200,  # sample size.
                     M = 3,   # true number of machines.
                     L = L0,   # number of antibody landmarks.
                     K = 8,    # number of true components.,
                     theta = rep(0.8,L0), # true positive rates
                     psi   = rep(0.01,L0), # false positive rates
                     alpha1 = 1 # half of the people have the first machine.
)

#image(simulate_data(options_sim0,SETSEED = TRUE)$datmat)
simu     <- simulate_data(options_sim0, SETSEED=TRUE)
simu_dat <- simu$datmat
simulate_Q_dat(5,simu_dat)
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14]
#> [1,]    0    0    0    0    0    0    0    0    0     0     0     0     0     0
#> [2,]    0    0    1    0    0    0    0    0    0     0     0     0     0     0
#> [3,]    0    0    0    0    0    0    0    0    0     0     0     0     0     0
#> [4,]    0    0    0    0    0    0    0    0    0     0     1     0     1     0
#> [5,]    0    0    0    0    0    0    0    0    0     0     1     0     0     0
#>      [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25] [,26]
#> [1,]     0     0     0     0     0     0     1     0     0     0     0     0
#> [2,]     0     0     0     0     1     0     0     0     0     0     0     0
#> [3,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [4,]     0     0     0     0     1     0     0     0     0     0     0     0
#> [5,]     0     0     0     0     0     0     0     0     0     0     0     0
#>      [,27] [,28] [,29] [,30] [,31] [,32] [,33] [,34] [,35] [,36] [,37] [,38]
#> [1,]     0     0     0     0     1     0     0     0     0     0     0     0
#> [2,]     0     0     0     0     0     0     0     0     0     0     0     1
#> [3,]     0     0     0     1     0     1     0     0     0     0     0     0
#> [4,]     0     0     1     0     0     1     0     0     0     0     0     0
#> [5,]     0     0     0     0     0     0     0     0     0     0     1     0
#>      [,39] [,40] [,41] [,42] [,43] [,44] [,45] [,46] [,47] [,48] [,49] [,50]
#> [1,]     0     0     1     0     0     0     0     0     0     0     0     0
#> [2,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [3,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [4,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [5,]     0     0     0     0     0     0     0     0     0     0     0     0
#>      [,51] [,52] [,53] [,54] [,55] [,56] [,57] [,58] [,59] [,60] [,61] [,62]
#> [1,]     0     0     0     1     0     0     0     0     1     0     0     0
#> [2,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [3,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [4,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [5,]     0     0     0     1     0     0     0     0     0     0     0     0
#>      [,63] [,64] [,65] [,66] [,67] [,68] [,69] [,70] [,71] [,72] [,73] [,74]
#> [1,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [2,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [3,]     1     0     0     0     0     0     0     0     0     0     0     0
#> [4,]     1     0     0     0     0     0     0     0     0     0     0     0
#> [5,]     0     0     0     0     0     0     0     1     0     0     0     0
#>      [,75] [,76] [,77] [,78] [,79] [,80] [,81] [,82] [,83] [,84] [,85] [,86]
#> [1,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [2,]     0     0     0     0     0     0     0     0     1     0     0     0
#> [3,]     0     0     0     0     1     0     0     0     1     0     0     0
#> [4,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [5,]     0     0     0     0     1     0     0     0     0     0     1     0
#>      [,87] [,88] [,89] [,90] [,91] [,92] [,93] [,94] [,95] [,96] [,97] [,98]
#> [1,]     0     0     0     0     1     0     0     0     0     0     0     0
#> [2,]     0     1     1     0     0     0     0     0     0     0     0     0
#> [3,]     0     0     0     0     0     0     0     0     0     1     0     0
#> [4,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [5,]     0     0     0     0     0     0     0     0     0     0     0     0
#>      [,99] [,100]
#> [1,]     0      0
#> [2,]     0      0
#> [3,]     0      0
#> [4,]     0      0
#> [5,]     0      0