Split-merge type update of clusterings

This function uses restricted_gibbs to arrive at a launch state and then propose a split-merge move. Based on the computed acceptance probability, this function accepts or rejects the proposal arriving at an old or new state for the cluster indicators.

split_merge(
  Y,
  z,
  zs,
  S,
  mylist,
  N,
  t,
  b,
  log_v,
  n_split,
  Q,
  p,
  theta,
  psi,
  MORE_SPLIT = NULL,
  partition_partial = NULL
)

Arguments

Y
multivariate binary data (row for subjects, column for dimensions)

z
original cluster indicators

zs
temporary variable used for split-merge assignments

S
temporary set of indicies for restricted Gibbs

mylist
the cluster ID list

N
a vector of the numbers of subjects in the list of clusters

t
total number of non-empty clusters

b
gamma parameter in the mixture of finite mixture formulation

log_v
coefficients in MFM (check Miller and Harrison 2016 JASA)

n_split
the number of intermediate Gibbs scan to arrive at the launch split state (usually 5)

Q
a Q matrix of dimension M by L

p
a vector of machine (factor) prevalences of length M

theta
a vector of true positive rates of length L

psi
a vector of false positive rates of length L

MORE_SPLIT
Default is NULL. When getting to launch state of the partition, TRUE for biasing towards split; FALSE for uniformly choose a pair (i,j) and then deciding to merge (if they belong to distinct clusters) or split (if they belong to the identical cluster)

partition_partial
a list of subject ids that each belong to a few known clusters.

Value

Returns the values at the end of the current iteration

t the number of clusters;

z the cluster assignment indicators for all subjects (taking values from mylist);

N the number of subjects for each cluster; please refer to mylist to get the matching cluster IDs;

mylist the non-empty and empty cluster IDs.