update true and false positive rates

This function samples the true (theta) and false (psi) positive rates; Constraint: theta[l]> 0.5 > psi[l]

update_positive_rate(Y, H, Q, a_theta, a_psi, old_theta_vec = NULL)

Arguments

Y
binary data matrix

H
a binary matrix: rows for all individuals, columns for latent states

Q
Q matrix

a_theta, a_psi
Beta hyperparameters for priors on theta and psi, respectively. 2 by L matrices.

old_theta_vec
sampled theta parameters from the previous iteration (vector of length L); to sample psi given the constraint

Value

a list of true theta and false psi positive rates, each of length L.